# Conflicts:
#	docs/US21/US21_IWantToGetTheNumberOfStudentsEnrolledInAProgrammeEdition_SequenceDiagram.puml

+++ src/main/java/PAI/controller/US28_RemoveTheEnrollmentOfAStudentInACourseEdition_Controller.java
@@ -1,16 +1,21 @@
 package PAI.controller;
 
-import PAI.domain.CourseEditionEnrollmentRepository;
+import PAI.repository.CourseEditionEnrollmentRepository;
 import PAI.domain.CourseEdition;
 import PAI.domain.Student;
 
 public class US28_RemoveTheEnrollmentOfAStudentInACourseEdition_Controller {
-    private CourseEditionEnrollmentRepository courseEditionEnrollmentRepository;
+
+    private final CourseEditionEnrollmentRepository courseEditionEnrollmentRepository;
 
     public US28_RemoveTheEnrollmentOfAStudentInACourseEdition_Controller(CourseEditionEnrollmentRepository courseEditionEnrollmentRepository) {
+
         this.courseEditionEnrollmentRepository = courseEditionEnrollmentRepository;
     }
-    public boolean removeStudentEnrolment(Student student, CourseEdition courseEdition) {
-        return courseEditionEnrollmentRepository.removeEnrollment(student, courseEdition);
+    public boolean removeStudentEnrollment(Student student, CourseEdition courseEdition) {
+        if(student==null || courseEdition==null){
+            throw new IllegalArgumentException("Student and CourseEdition cannot be null");
+        }
+           return courseEditionEnrollmentRepository.removeEnrollment(student, courseEdition);
     }
 }
